Separate home logic to custom hook useCart
